64-bit fixes
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@152163 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/sqlite/sqlite.c
@@ -1823,7 +1823,7 @@ PHP_FUNCTION(sqlite_fetch_all)
 PHP_FUNCTION(sqlite_fetch_array)
 {
 	zval *zres;
-	int mode = PHPSQLITE_BOTH;
+	long mode = PHPSQLITE_BOTH;
 	zend_bool decode_binary = 1;
 	struct php_sqlite_result *res;
 	zval *object = getThis();
